requirements.txt
flask
webexteamssdk
openai
requests
python-dotenv
